requirements.txt
fastapi
uvicorn
joblib
xgboost
scikit-learn
pandas
numpy
requests
python-dotenv